requirements/base.txt
# Bleeding edge Django
django>=1.8.7,<1.9.0

# Configuration
django-environ>=0.4.0
django-secure>=1.0.1
whitenoise>=2.0.4


# Forms
django-braces>=1.8.1
django-floppyforms>=1.6.1

# Models
django-model-utils>=2.3.1

# Images
Pillow==3.0.0

django-userena

# Python-PostgreSQL Database Adapter
#psycopg2==2.6.1

# Unicode slugification
unicode-slugify>=0.1.3
django-autoslug>=1.9.3

# Time zones support
pytz==2015.7

# Redis support
django-redis>=4.3.0
redis>=2.10.0

kombu
celery>=3.1.19

# Static and Media Storage
# ------------------------------------------------
django-storages-redux>=1.3

# Your custom requirements go here

django-vanilla-views
django-js-reverse
django-hijack
pathlib
django-ratelimit
bcrypt
PyMySQL
django-flat-theme
django-easy-timezones
